Support Measure return for FTQC
We want the runtime mz to return a boolean in FTQC. For NISQ, the boolean return is just a dummy one.

Whenever we need to use the return value of the Measure (e.g. assignment or conditional), it's parsed as a classical instruction.

Hence, we just need to do token replacement in that code path.

Also, fixing the `else` path of the token collector to prevent it from combining the next line (could be a quantum instruction) into the `else` line.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>