Support KernelSignature in vector
With KernelSignature in a vector, we need to be able to reliably add the parent_kernel to the operator() call.
It's error-prone to try doing that at the token collector level, hence using a type-safe helper to iterate over the kernel arguments.
We'll pick up any KernelSignature vars and attach the proper parent_kernel => i.e. its operator() invocation w/o the CompositeInstruction arg will be using the correct parent composite.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>